requirements.txt
PyYAML==3.10
argh==0.24.1
globusonline-transfer-api-client==0.10.16
mercurial==2.8.2
nose==1.3.0
pathtools==0.1.2
watchdog==0.7.1
wsgiref==0.1.2
